Move ctuple tests to separate file
